disable tfo if initial send fails
In general, we only send addr buffer in the initial
ConnectAsync, it should not exceed MSS
         IPv4    Domain    IPv6
Origin  1+4+2, 1+1+255+2, 1+16+2
AEAD    length above plus AEAD overhead and garbage

If something went wrong, e.g. not all data are being
sent out, we try to disable tfo and try the
traditional 3-way handshaking in the subsequent connections.

This situation may happen when it contains many garbage.